requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-guardian==2.1.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rules==2.2
sqlparse==0.3.0
urllib3==1.25.8
